DEV RELEASE OF VERV-76

+++ internal/config/custom_extractors.go
@@ -1,10 +1,10 @@
 package config
 
-func GetAvailablePorts(cfg Config) ([]uint16, error) {
+func GetAvailablePorts(cfg Config) ([]uint32, error) {
 	ap := cfg.GetEnvironment().AvailablePorts
-	out := make([]uint16, 0, len(ap))
+	out := make([]uint32, 0, len(ap))
 	for _, p := range ap {
-		out = append(out, uint16(p))
+		out = append(out, uint32(p))
 	}
 
 	return out, nil
